requirements.txt
pandas
playwright
google-api-python-client
google-auth
python-dotenv
